+++ spec/lumberjack/log_entry_spec.rb
@@ -57,6 +57,19 @@ RSpec.describe Lumberjack::LogEntry do
     expect(entry.tags).to eq("unit_of_work_id" => "ABCD")
   end
 
+  it "should compact tags that are set to empty values" do
+    tags = {
+      "a" => "A",
+      "b" => nil,
+      "c" => "",
+      "d" => {"e" => "E", "f" => nil},
+      "g" => {"h" => "", "i" => []}
+    }
+
+    entry = Lumberjack::LogEntry.new(Time.now, Logger::INFO, "test", "app", 1500, tags)
+    expect(entry.tags).to eq("a" => "A", "d" => {"e" => "E"})
+  end
+
   it "should have a unit_of_work_id for backward compatibility with the 1.0 API", suppress_warnings: true do
     entry = Lumberjack::LogEntry.new(Time.now, Logger::INFO, "test", "app", 1500, "ABCD")
     expect(entry.unit_of_work_id).to eq("ABCD")